+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+-   New "List Users By Course" sub-section to "Courses".
+-   New "List Courses By User" sub-section to "Users".
+
+## [0.0.3] - 2021-01-18
+
+### Added
+
+-   New "Bulk User Unenrollment" sub-section to "Enrollment".
+-   New "List Users" AJAX request route in "Courses".
+-   New "List Courses" AJAX request route in "Users".
+
 ## [0.0.2] - 2021-01-17
 
 ### Added
@@ -28,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _Initial major release.._
 
-[unreleased]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.2...HEAD
+[unreleased]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.3...HEAD
+[0.0.3]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/ManuelGil/alternate-admin/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/ManuelGil/alternate-admin/releases/tag/v0.0.1